Python/requirements.txt
gradio
torch
torchvision
transformers
numpy
Pillow
huggingface_hub
regex
accelerate
bitsandbytes
python-dotenv
requests